requirements.txt
yfinance>=0.2.0
pandas>=1.5.0
openpyxl>=3.0.0
akshare>=1.12.0
baostock>=0.8.0
efinance>=0.5.0
requests>=2.28.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.0
